+++ tests/forms/test_xso.py
@@ -460,6 +460,16 @@ class TestField(unittest.TestCase):
                 "duplicate option label"):
             f.validate()
 
+    def test_allow_duplicate_Nones_in_option_labels(self):
+        f = forms_xso.Field()
+        f.type_ = forms_xso.FieldType.LIST_MULTI
+        f.var = "foobar"
+        f.options["foo"] = "bar"
+        f.options["baz"] = None
+        f.options["fnord"] = None
+
+        f.validate()
+
     def test_init(self):
         f = forms_xso.Field()
         self.assertEqual(f.type_, forms_xso.FieldType.TEXT_SINGLE)